Create first version
This commit does a ton -- it should be broken into smaller pieces but
since this is a quick project I'm going to be bad and commit it all at
once.

I'm using polyfill as an attempt to find the center of my polygons.

The decision to use Vue and TypeScript is really just because I wanted
to practice both of them.